+++ common/protocol/src/methods/helpers/archiveDebugBundle.ts
@@ -1,10 +1,16 @@
 import { existsSync, mkdirSync, createWriteStream, readFileSync } from "fs";
 import path from "path";
+<<<<<<< HEAD
 import { standardizeError, Validator } from "../..";
 import JSZip from "jszip";
 import { VoteType } from "@kyvejs/types/client/kyve/bundles/v1beta1/tx";
 import { DataItem } from "../../proto/kyverdk/runtime/v1/runtime";
 import Diff from "diff";
+=======
+import { DataItem, dirSize, standardizeError, Validator } from "../..";
+import JSZip from "jszip";
+import { VoteType } from "@kyvejs/types/client/kyve/bundles/v1beta1/tx";
+>>>>>>> main
 
 /**
  * archiveDebugBundle is used for storing a bundle for debug
@@ -26,76 +32,91 @@ export function archiveDebugBundle(
   metadata: object
 ): void {
   try {
-    // this.logger.info("Archiving debug bundle");
-    // // if "debug" folder under target path does not exist create it
-    // if (!existsSync(path.join(this.home, `debug`))) {
-    //   mkdirSync(path.join(this.home, `debug`), { recursive: true });
-    // }
-    // const zip = new JSZip();
-    // // save metadata which includes vote reasons and args
-    // const metadata_str = JSON.stringify(metadata || {}, null, 2);
-    // zip.file("metadata.json", metadata_str);
-    // // save current pool state including the raw bundle proposal
-    // const pool_str = JSON.stringify(this.pool || {}, null, 2);
-    // zip.file("pool.json", pool_str);
-    // // save the proposed bundle from the uploader
-    // const proposed_bundle_str = JSON.stringify(proposedBundle || [], null, 2);
-    // zip.file("proposed_bundle.json", proposed_bundle_str);
-    // // save the locally created bundle from this node
-    // const validation_bundle_str = JSON.stringify(
-    //   validationBundle || [],
-    //   null,
-    //   2
-    // );
-    // zip.file("validation_bundle.json", validation_bundle_str);
-    // // save the diff of the proposed and local bundle
-    // const diff_str = Diff.createTwoFilesPatch(
-    //   "proposed_bundle.json",
-    //   "validation_bundle.json",
-    //   proposed_bundle_str,
-    //   validation_bundle_str
-    // );
-    // zip.file("diff.txt", diff_str);
-    // // save the logfile of the current session
-    // const debug_str = readFileSync(path.join(this.home, "logs", this.logFile));
-    // zip.file("debug.log", debug_str);
-    // // get human readable vote
-    // let voteType = "";
-    // switch (vote) {
-    //   case VoteType.VOTE_TYPE_VALID:
-    //     voteType = "valid";
-    //     break;
-    //   case VoteType.VOTE_TYPE_INVALID:
-    //     voteType = "invalid";
-    //     break;
-    //   case VoteType.VOTE_TYPE_ABSTAIN:
-    //     voteType = "abstain";
-    //     break;
-    //   case VoteType.VOTE_TYPE_UNSPECIFIED:
-    //     voteType = "unspecified";
-    //     break;
-    //   default:
-    //     voteType = "unrecognized";
-    // }
-    // const storageId = this.pool?.bundle_proposal?.storage_id ?? "";
-    // const zipPath = path.join(
-    //   this.home,
-    //   `debug`,
-    //   `${voteType}_${this.pool.id}_${Math.floor(
-    //     Date.now() / 1000
-    //   )}_${storageId.slice(0, 6)}.zip`
-    // );
-    // // save zip file
-    // zip
-    //   .generateNodeStream({ type: "nodebuffer", streamFiles: true })
-    //   .pipe(createWriteStream(zipPath))
-    //   .on("finish", () => {
-    //     this.logger.info("Successfully saved debug information");
-    //   })
-    //   .on("error", (err) => {
-    //     this.logger.error("Failed to save debug information");
-    //     this.logger.error(standardizeError(err));
-    //   });
+    this.logger.info("Archiving debug bundle");
+
+    // if "debug" folder under target path does not exist create it
+    if (!existsSync(path.join(this.home, `debug`))) {
+      mkdirSync(path.join(this.home, `debug`), { recursive: true });
+    }
+
+    // if size of debug folder exceeds debug max limit we don't store an archive
+    if (this.debugMaxSize > 0) {
+      const debugDirSize = dirSize(path.join(this.home, `debug`));
+
+      if (debugDirSize >= this.debugMaxSize) {
+        this.logger.warn(
+          `Skipped saving debug information, debug dir exceeded max size of ${this.debugMaxSize} with ${debugDirSize} bytes`
+        );
+        return;
+      }
+    }
+
+    const zip = new JSZip();
+
+    // save metadata which includes vote reasons and args
+    const metadata_str = JSON.stringify(metadata || {}, null, 2);
+    zip.file("metadata.json", metadata_str);
+
+    // save current pool state including the raw bundle proposal
+    const pool_str = JSON.stringify(this.pool || {}, null, 2);
+    zip.file("pool.json", pool_str);
+
+    // save the proposed bundle from the uploader
+    const proposed_bundle_str = JSON.stringify(proposedBundle || [], null, 2);
+    zip.file("proposed_bundle.json", proposed_bundle_str);
+
+    // save the locally created bundle from this node
+    const validation_bundle_str = JSON.stringify(
+      validationBundle || [],
+      null,
+      2
+    );
+    zip.file("validation_bundle.json", validation_bundle_str);
+
+    // save the logfile of the current session
+    const debug_str = readFileSync(path.join(this.home, "logs", this.logFile));
+    zip.file("debug.log", debug_str);
+
+    // get human readable vote
+    let voteType = "";
+
+    switch (vote) {
+      case VoteType.VOTE_TYPE_VALID:
+        voteType = "valid";
+        break;
+      case VoteType.VOTE_TYPE_INVALID:
+        voteType = "invalid";
+        break;
+      case VoteType.VOTE_TYPE_ABSTAIN:
+        voteType = "abstain";
+        break;
+      case VoteType.VOTE_TYPE_UNSPECIFIED:
+        voteType = "unspecified";
+        break;
+      default:
+        voteType = "unrecognized";
+    }
+
+    const storageId = this.pool?.bundle_proposal?.storage_id ?? "";
+    const zipPath = path.join(
+      this.home,
+      `debug`,
+      `${voteType}_${this.pool.id}_${Math.floor(
+        Date.now() / 1000
+      )}_${storageId.slice(0, 6)}.zip`
+    );
+
+    // save zip file
+    zip
+      .generateNodeStream({ type: "nodebuffer", streamFiles: true })
+      .pipe(createWriteStream(zipPath))
+      .on("finish", () => {
+        this.logger.info("Successfully saved debug information");
+      })
+      .on("error", (err) => {
+        this.logger.error("Failed to save debug information");
+        this.logger.error(standardizeError(err));
+      });
   } catch (err) {
     this.logger.error("Failed to save debug information");
     this.logger.error(standardizeError(err));