+++ common/protocol/test/propose_bundle.test.ts
@@ -25,6 +25,7 @@ TEST CASES - propose bundle tests
 * propose bundle where saveBundle and skipUploaderRole fails
 * propose bundle where summarizeDataBundle fails
 * propose bundle where compress fails
+* propose bundle with the same data as last round
 
 */
 
@@ -260,6 +261,12 @@ describe("propose bundle tests", () => {
     // ASSERT NODEJS INTERFACES
     // ========================
 
+    expect(v["lastUploadedBundle"]).toEqual({
+      storageId: "test_storage_id",
+      dataHash: sha256(Buffer.from(JSON.stringify(bundle))),
+      dataSize: Buffer.from(JSON.stringify(bundle)).byteLength,
+    });
+
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
 
@@ -419,6 +426,8 @@ describe("propose bundle tests", () => {
     // ASSERT NODEJS INTERFACES
     // ========================
 
+    expect(v["lastUploadedBundle"]).toEqual(null);
+
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
 
@@ -553,6 +562,8 @@ describe("propose bundle tests", () => {
     // ========================
     // ASSERT NODEJS INTERFACES
     // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual(null);
 
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
@@ -701,6 +712,12 @@ describe("propose bundle tests", () => {
     // ========================
     // ASSERT NODEJS INTERFACES
     // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual({
+      storageId: "test_storage_id",
+      dataHash: sha256(Buffer.from(JSON.stringify(bundle))),
+      dataSize: Buffer.from(JSON.stringify(bundle)).byteLength,
+    });
 
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
@@ -852,6 +869,8 @@ describe("propose bundle tests", () => {
     // ASSERT NODEJS INTERFACES
     // ========================
 
+    expect(v["lastUploadedBundle"]).toEqual(null);
+
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
 
@@ -1001,6 +1020,8 @@ describe("propose bundle tests", () => {
     // ========================
     // ASSERT NODEJS INTERFACES
     // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual(null);
 
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
@@ -1161,6 +1182,12 @@ describe("propose bundle tests", () => {
     // ASSERT NODEJS INTERFACES
     // ========================
 
+    expect(v["lastUploadedBundle"]).toEqual({
+      storageId: "test_storage_id",
+      dataHash: sha256(Buffer.from(JSON.stringify(bundle))),
+      dataSize: Buffer.from(JSON.stringify(bundle)).byteLength,
+    });
+
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
 
@@ -1299,6 +1326,8 @@ describe("propose bundle tests", () => {
     // ========================
     // ASSERT NODEJS INTERFACES
     // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual(null);
 
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
@@ -1454,6 +1483,8 @@ describe("propose bundle tests", () => {
     // ASSERT NODEJS INTERFACES
     // ========================
 
+    expect(v["lastUploadedBundle"]).toEqual(null);
+
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
 
@@ -1596,6 +1627,8 @@ describe("propose bundle tests", () => {
     // ========================
     // ASSERT NODEJS INTERFACES
     // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual(null);
 
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
@@ -1742,6 +1775,184 @@ describe("propose bundle tests", () => {
     // ========================
     // ASSERT NODEJS INTERFACES
     // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual(null);
+
+    // assert that only one round ran
+    expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);
+
+    // TODO: assert timeouts
+  });
+
+  test("propose bundle with the same data as last round", async () => {
+    // ARRANGE
+    const bundle = [
+      {
+        key: "test_key_1",
+        value: "test_value_1",
+      },
+      {
+        key: "test_key_2",
+        value: "test_value_2",
+      },
+      {
+        key: "test_key_3",
+        value: "test_value_3",
+      },
+      {
+        key: "test_key_4",
+        value: "test_value_4",
+      },
+    ];
+
+    v["lastUploadedBundle"] = {
+      storageId: "last_storage_id",
+      dataHash: sha256(Buffer.from(JSON.stringify(bundle))),
+      dataSize: Buffer.from(JSON.stringify(bundle)).byteLength,
+    };
+
+    v["lcd"][0].kyve.query.v1beta1.canVote = jest.fn().mockResolvedValue({
+      possible: false,
+      reason: "Already voted",
+    });
+
+    v["syncPoolState"] = jest.fn().mockImplementation(() => {
+      v.pool = {
+        ...genesis_pool,
+        data: {
+          ...genesis_pool.data,
+          current_index: "100",
+          current_key: "99",
+        },
+        bundle_proposal: {
+          ...genesis_pool.bundle_proposal,
+          storage_id: "another_test_storage_id",
+          uploader: "another_test_staker",
+          next_uploader: "test_staker",
+          data_size: "123456789",
+          data_hash: "previous_test_bundle_hash",
+          bundle_size: "2",
+          from_key: "100",
+          to_key: "101",
+          bundle_summary: "previous_test_value",
+          updated_at: "0",
+          voters_valid: ["test_staker"],
+        },
+      } as any;
+    });
+
+    await v["cacheProvider"].put("102", bundle[0]);
+    await v["cacheProvider"].put("103", bundle[1]);
+    await v["cacheProvider"].put("104", bundle[2]);
+    await v["cacheProvider"].put("105", bundle[3]);
+
+    // ACT
+    await runNode.call(v);
+
+    // ASSERT
+    const txs = v["client"][0].kyve.bundles.v1beta1;
+    const queries = v["lcd"][0].kyve.query.v1beta1;
+    const cacheProvider = v["cacheProvider"];
+    const runtime = v["runtime"];
+
+    // ========================
+    // ASSERT CLIENT INTERFACES
+    // ========================
+
+    expect(txs.claimUploaderRole).toHaveBeenCalledTimes(0);
+
+    expect(txs.voteBundleProposal).toHaveBeenCalledTimes(0);
+
+    expect(txs.submitBundleProposal).toHaveBeenCalledTimes(1);
+    expect(txs.submitBundleProposal).toHaveBeenLastCalledWith({
+      staker: "test_staker",
+      pool_id: "0",
+      storage_id: "last_storage_id",
+      data_size: Buffer.from(JSON.stringify(bundle)).byteLength.toString(),
+      data_hash: sha256(Buffer.from(JSON.stringify(bundle))),
+      from_index: "102",
+      bundle_size: "4",
+      from_key: "test_key_1",
+      to_key: "test_key_4",
+      bundle_summary: JSON.stringify(bundle),
+    });
+
+    expect(txs.skipUploaderRole).toHaveBeenCalledTimes(0);
+
+    // =====================
+    // ASSERT LCD INTERFACES
+    // =====================
+
+    expect(queries.canVote).toHaveBeenCalledTimes(1);
+    expect(queries.canVote).toHaveBeenLastCalledWith({
+      staker: "test_staker",
+      pool_id: "0",
+      voter: "test_valaddress",
+      storage_id: "another_test_storage_id",
+    });
+
+    expect(queries.canPropose).toHaveBeenCalledTimes(1);
+    expect(queries.canPropose).toHaveBeenLastCalledWith({
+      staker: "test_staker",
+      pool_id: "0",
+      proposer: "test_valaddress",
+      from_index: "102",
+    });
+
+    // =========================
+    // ASSERT STORAGE INTERFACES
+    // =========================
+
+    expect(storageProvider.saveBundle).toHaveBeenCalledTimes(0);
+    expect(storageProvider.retrieveBundle).toHaveBeenCalledTimes(0);
+
+    // =======================
+    // ASSERT CACHE INTERFACES
+    // =======================
+
+    expect(cacheProvider.get).toHaveBeenCalledTimes(5);
+    expect(cacheProvider.get).toHaveBeenNthCalledWith(1, "102");
+    expect(cacheProvider.get).toHaveBeenNthCalledWith(2, "103");
+    expect(cacheProvider.get).toHaveBeenNthCalledWith(3, "104");
+    expect(cacheProvider.get).toHaveBeenNthCalledWith(4, "105");
+    expect(cacheProvider.get).toHaveBeenNthCalledWith(5, "106");
+
+    // =============================
+    // ASSERT COMPRESSION INTERFACES
+    // =============================
+
+    expect(compression.compress).toHaveBeenCalledTimes(1);
+    expect(compression.compress).toHaveBeenLastCalledWith(
+      Buffer.from(JSON.stringify(bundle))
+    );
+
+    expect(compression.decompress).toHaveBeenCalledTimes(0);
+
+    // =========================
+    // ASSERT RUNTIME INTERFACES
+    // =========================
+
+    expect(runtime.summarizeDataBundle).toHaveBeenCalledTimes(1);
+    expect(runtime.summarizeDataBundle).toHaveBeenLastCalledWith(
+      expect.any(Validator),
+      bundle
+    );
+
+    expect(runtime.validateDataItem).toHaveBeenCalledTimes(0);
+
+    expect(runtime.getDataItem).toHaveBeenCalledTimes(0);
+
+    expect(runtime.nextKey).toHaveBeenCalledTimes(0);
+
+    // ========================
+    // ASSERT NODEJS INTERFACES
+    // ========================
+
+    expect(v["lastUploadedBundle"]).toEqual({
+      storageId: "last_storage_id",
+      dataHash: sha256(Buffer.from(JSON.stringify(bundle))),
+      dataSize: Buffer.from(JSON.stringify(bundle)).byteLength,
+    });
 
     // assert that only one round ran
     expect(v["waitForNextBundleProposal"]).toHaveBeenCalledTimes(1);